Implement sort. Write output to file.
The toolchain currently uses paths to get access time and size. It
returns tuples, which are added to a list, and later sorted, and finally
written to file.

This is an MVP. Next Step TODO: make things easier to sort through. Tack
on a frontend to this bullshit. Pick an appropriate data structure for
this.